fix: name consistent with spec and test files
