+++ tests/test_all.cc
@@ -5,11 +5,14 @@
 #include "../src/nyx/globals.h"
 #include "test_contour.h"
 #include "test_firstorder_regression.h"
+#include "test_firstorder_analytic.h"
+#include "test_firstorder_matlab.h"
 #include "test_intensity_histogram_regression.h"
+#include "test_intensity_histogram_analytic.h"
 #include "test_intensity_histogram_ibsi.h"
 #include "test_hu_analytic.h"
 #include "test_hu_mechanics.h"
-#include "test_morphology_features.h"
+#include "test_morphology_matlab_perimeter.h"
 #include "test_morphology_regression.h"
 #include "test_morphology_analytic.h"
 #include "test_morphology_imea.h"
@@ -33,6 +36,7 @@
 #include "test_firstorder_pyradiomics.h"
 #include "test_ngldm_ibsi.h"
 #include "test_ngtdm_ibsi.h"
+#include "test_glcm_pyradiomics.h"
 #include "test_glcm_regression.h"
 #include "test_gldm_regression.h"
 #include "test_gldm_mechanics.h"
@@ -49,7 +53,7 @@
 #include "test_3d_morphology_common.h"
 #include "test_3d_morphology_regression.h"
 #include "test_3d_morphology_matlab.h"
-#include "test_3d_gldzm_ibsi.h"
+#include "test_3d_gldzm_regression.h"
 #include "test_3d_ngldm_regression.h"
 #include "test_3d_firstorder_pyradiomics.h"
 #include "test_3d_glcm_pyradiomics.h"
@@ -57,16 +61,13 @@
 #include "test_3d_ngtdm_pyradiomics.h"
 #include "test_3d_glrlm_pyradiomics.h"
 #include "test_3d_glszm_pyradiomics.h"
-#include "test_3d_coverage_common.h"
-#include "test_3d_firstorder_coverage.h"
-#include "test_3d_morphology_coverage.h"
-#include "test_3d_glcm_coverage.h"
-#include "test_3d_gldm_coverage.h"
-#include "test_3d_gldzm_coverage.h"
-#include "test_3d_glrlm_coverage.h"
-#include "test_3d_glszm_coverage.h"
-#include "test_3d_ngldm_coverage.h"
-#include "test_3d_ngtdm_coverage.h"
+#include "test_3d_regression_common.h"
+#include "test_3d_firstorder_regression.h"
+#include "test_3d_glcm_regression.h"
+#include "test_3d_gldm_regression.h"
+#include "test_3d_glrlm_regression.h"
+#include "test_3d_glszm_regression.h"
+#include "test_3d_ngtdm_regression.h"
 #include "test_glcm_mechanics.h"
 #ifdef USE_ARROW
     #include "test_arrow_mechanics.h"
