+++ tests/test_firstorder_regression.h
@@ -9,15 +9,13 @@
 #include "test_data.h"
 #include "test_main_nyxus.h"
 
+// NOTE: Oracle-claiming first-order assertions live in
+// test_firstorder_analytic.h and test_firstorder_matlab.h (SPEC §2/#6).
+// This file is regression / snapshot only.
+
 using namespace Nyxus;
 
 // ROI pixel accumulation routines implemented in Nyxus
-static constexpr double oracle_3p_builtin_hyperskewness_feature_golden_value = 1.978293086605381;
-static constexpr double oracle_3p_builtin_hyperflatness_feature_golden_value = 5.126659243028459;
-static constexpr double oracle_3p_matlab_uniformity_feature_golden_value = 0.0647664;
-static constexpr double oracle_3p_builtin_uniformity_piu_feature_golden_value = 29.477577192725725;
-static constexpr double oracle_3p_builtin_covered_image_intensity_range_feature_golden_value = 8.088960097657740e-01;
-static constexpr double oracle_3p_builtin_robust_mean_feature_golden_value = 3.142136800000000e+04;
 
 static void calculate_pixel_intensity_feature_values(
     std::vector<std::vector<double>>& fvals,
@@ -254,62 +252,6 @@ void test_pixel_intensity_pearson_kurtosis()
     ASSERT_TRUE(agrees_gt(fvals[(int)Nyxus::Feature2D::KURTOSIS][0], 1.927888720710090));
 }
 
-void test_pixel_intensity_verifiable_with_3p_builtin_oracle_hyperskewness()
-{
-    SCOPED_TRACE("VERIFIABLE_WITH_3P_BUILTIN_ORACLE__HYPERSKEWNESS");
-
-    // Feed data to the ROI
-    Dataset ds;
-    ds.dataset_props.push_back(SlideProps("",""));
-
-    LR roidata(100);   // dummy label 100
-    roidata.slide_idx = -1; // we don't have a real slide for this test ROI
-    load_test_roi_data(roidata, pixelIntensityFeaturesTestData, sizeof(pixelIntensityFeaturesTestData) / sizeof(NyxusPixel));
-
-    // Anisotropy (none)
-    roidata.make_nonanisotropic_aabb();
-
-    // Calculate features
-    PixelIntensityFeatures f;
-    Fsettings s;
-    ASSERT_NO_THROW(f.calculate(roidata, s, ds));
-
-    // Retrieve the feature values
-    roidata.initialize_fvals();
-    f.save_value(roidata.fvals);
-
-    // Check the feature values vs ground truth
-    ASSERT_TRUE(agrees_gt(roidata.fvals[(int)Nyxus::Feature2D::HYPERSKEWNESS][0], oracle_3p_builtin_hyperskewness_feature_golden_value));
-}
-
-void test_pixel_intensity_verifiable_with_3p_builtin_oracle_hyperflatness()
-{
-    SCOPED_TRACE("VERIFIABLE_WITH_3P_BUILTIN_ORACLE__HYPERFLATNESS");
-
-    // Feed data to the ROI
-    Dataset ds;
-    ds.dataset_props.push_back(SlideProps("",""));
-
-    LR roidata(100);   // dummy label 100
-    roidata.slide_idx = -1; // we don't have a real slide for this test ROI
-    load_test_roi_data(roidata, pixelIntensityFeaturesTestData, sizeof(pixelIntensityFeaturesTestData) / sizeof(NyxusPixel));
-
-    // Anisotropy (none)
-    roidata.make_nonanisotropic_aabb();
-
-    // Calculate features
-    PixelIntensityFeatures f;
-    Fsettings s;
-    ASSERT_NO_THROW(f.calculate(roidata, s, ds));
-
-    // Retrieve the feature values
-    roidata.initialize_fvals();
-    f.save_value(roidata.fvals);
-
-    // Check the feature values vs ground truth
-    ASSERT_TRUE(agrees_gt(roidata.fvals[(int)Nyxus::Feature2D::HYPERFLATNESS][0], oracle_3p_builtin_hyperflatness_feature_golden_value));
-}
-
 void test_pixel_intensity_mean_absolute_deviation()
 {
     // Feed data to the ROI
@@ -454,69 +396,6 @@ void test_pixel_intensity_energy()
     ASSERT_TRUE(agrees_gt(roidata.fvals[(int)Nyxus::Feature2D::ENERGY][0], 1.965289571840000e+11));
 }
 
-void test_pixel_intensity_uniformity()
-{
-    // Feed data to the ROI
-    Dataset ds;
-    ds.dataset_props.push_back(SlideProps("",""));
-
-    LR roidata(100);   // dummy label 100
-    roidata.slide_idx = -1; // we don't have a real slide for this test ROI
-    load_test_roi_data(roidata, pixelIntensityFeaturesTestData, sizeof(pixelIntensityFeaturesTestData) / sizeof(NyxusPixel));
-
-    // Anisotropy (none)
-    roidata.make_nonanisotropic_aabb();
-
-    // settings important for this feature
-    Fsettings s;
-    s.resize((int)NyxSetting::__COUNT__);
-    s[(int)NyxSetting::GREYDEPTH].ival = 20;
-    s[(int)NyxSetting::USEGPU].bval = false;
-    s[(int)NyxSetting::IBSI].bval = false;
-
-    // Calculate features
-    PixelIntensityFeatures f;
-    ASSERT_NO_THROW(f.calculate(roidata, s, ds));
-
-    // Retrieve the feature values
-    roidata.initialize_fvals();
-    f.save_value(roidata.fvals);
-
-    // Check the feature values vs ground truth
-    ASSERT_TRUE(agrees_gt(roidata.fvals[(int)Nyxus::Feature2D::UNIFORMITY][0], oracle_3p_matlab_uniformity_feature_golden_value, 100)); // Using 1% tolerance vs MATLAB
-}
-
-void test_pixel_intensity_verifiable_with_3p_builtin_oracle_uniformity_piu()
-{
-    SCOPED_TRACE("VERIFIABLE_WITH_3P_BUILTIN_ORACLE__UNIFORMITY_PIU");
-
-    // Feed data to the ROI
-    Dataset ds;
-    ds.dataset_props.push_back(SlideProps("",""));
-
-    LR roidata(100);   // dummy label 100
-    roidata.slide_idx = -1; // we don't have a real slide for this test ROI
-    load_test_roi_data(roidata, pixelIntensityFeaturesTestData, sizeof(pixelIntensityFeaturesTestData) / sizeof(NyxusPixel));
-
-    // settings important for this feature
-    Fsettings s;
-    s.resize((int)NyxSetting::__COUNT__);
-    s[(int)NyxSetting::GREYDEPTH].ival = 20;
-    s[(int)NyxSetting::USEGPU].bval = false;
-    s[(int)NyxSetting::IBSI].bval = false;
-
-    // Calculate features
-    PixelIntensityFeatures f;
-    ASSERT_NO_THROW(f.calculate(roidata, s, ds));
-
-    // Retrieve the feature values
-    roidata.initialize_fvals();
-    f.save_value(roidata.fvals);
-
-    // Check the feature values vs ground truth
-    ASSERT_TRUE(agrees_gt(roidata.fvals[(int)Nyxus::Feature2D::UNIFORMITY_PIU][0], oracle_3p_builtin_uniformity_piu_feature_golden_value));
-}
-
 void test_pixel_intensity_percentiles_iqr()
 {
     std::vector<std::vector<double>> fvals;
@@ -539,16 +418,6 @@ void test_pixel_intensity_cov()
     ASSERT_TRUE(agrees_gt(fvals[(int)Nyxus::Feature2D::COV][0], 4.523365498399634e-01));
 }
 
-void test_pixel_intensity_verifiable_with_3p_builtin_oracle_covered_image_intensity_range()
-{
-    SCOPED_TRACE("VERIFIABLE_WITH_3P_BUILTIN_ORACLE__COVERED_IMAGE_INTENSITY_RANGE");
-
-    std::vector<std::vector<double>> fvals;
-    calculate_pixel_intensity_feature_values(fvals, Fsettings(), 0, 0.0, 65535.0);
-
-    ASSERT_TRUE(agrees_gt(fvals[(int)Nyxus::Feature2D::COVERED_IMAGE_INTENSITY_RANGE][0], oracle_3p_builtin_covered_image_intensity_range_feature_golden_value));
-}
-
 void test_pixel_intensity_median_absolute_deviation()
 {
     std::vector<std::vector<double>> fvals;
@@ -563,16 +432,6 @@ void test_pixel_intensity_qcod()
     calculate_pixel_intensity_feature_values(fvals);
 
     ASSERT_TRUE(agrees_gt(fvals[(int)Nyxus::Feature2D::QCOD][0], 4.119607630640470e-01));
-}
-
-void test_pixel_intensity_verifiable_with_3p_builtin_oracle_robust_mean()
-{
-    SCOPED_TRACE("VERIFIABLE_WITH_3P_BUILTIN_ORACLE__ROBUST_MEAN");
-
-    std::vector<std::vector<double>> fvals;
-    calculate_pixel_intensity_feature_values(fvals);
-
-    ASSERT_TRUE(agrees_gt(fvals[(int)Nyxus::Feature2D::ROBUST_MEAN][0], oracle_3p_builtin_robust_mean_feature_golden_value));
 }
 
 void test_pixel_intensity_standard_deviation_biased()
