+++ tests/test_glcm_mechanics.h
@@ -14,7 +14,7 @@
 // settings and left the GLCM-specific ones (GLCM_OFFSET / GLCM_GREYDEPTH / GLCM_NUMANG)
 // zero-initialised. With GLCM_OFFSET == 0 the co-occurrence shift is dx = dy = 0, so every
 // pixel co-occurs with itself -> a purely diagonal matrix -> CONTRAST = 0, CORRELATION = 1 for
-// any image. The existing test_glcm.h cases never caught this because they hard-code
+// any image. The existing test_glcm_regression.h cases never caught this because they hard-code
 // GLCM_OFFSET = 1 (and run on a fully-masked phantom). This test exercises the real default.
 inline void test_glcm_bug_offset_default_is_one()
 {
