+++ tests/vetting/coverage_report.md
@@ -2,24 +2,24 @@
 
 _Generated by check_coverage.py from oracle_coverage.csv. Do not edit by hand._
 
-Features vetted by >=1 oracle: 619/758 (82%)
-regression: 139  untested: 0
+Features vetted by >=1 oracle: 471/758 (62%)
+regression: 287  untested: 0
 
 | family | total | vetted | regression | untested |
 |---|---|---|---|---|
-| firstorder | 72 | 71 | 1 | 0 |
+| firstorder | 72 | 41 | 31 | 0 |
 | gabor | 1 | 1 | 0 | 0 |
-| glcm | 118 | 118 | 0 | 0 |
+| glcm | 118 | 92 | 26 | 0 |
 | gldm | 28 | 28 | 0 | 0 |
 | gldzm | 36 | 18 | 18 | 0 |
-| glrlm | 64 | 64 | 0 | 0 |
+| glrlm | 64 | 54 | 10 | 0 |
 | glszm | 32 | 32 | 0 | 0 |
 | imq | 6 | 4 | 2 | 0 |
 | intensity_histogram | 47 | 47 | 0 | 0 |
-| moments | 180 | 118 | 62 | 0 |
-| morphology | 113 | 81 | 32 | 0 |
+| moments | 180 | 86 | 94 | 0 |
+| morphology | 113 | 31 | 82 | 0 |
 | neighbor | 9 | 8 | 1 | 0 |
-| ngldm | 38 | 20 | 18 | 0 |
+| ngldm | 38 | 19 | 19 | 0 |
 | ngtdm | 10 | 10 | 0 | 0 |
 | radial | 3 | 0 | 3 | 0 |
 | zernike | 1 | 0 | 1 | 0 |
